[backend] refactor(registration): move files to their own module
