+++ apps/portal-api/src/modules/registration/registration.deprecated.graphql
