+++ apps/portal-api/src/modules/registration/registration.domain.ts
@@ -1,6 +1,6 @@
 import { Knex } from 'knex';
 import { v4 as uuidv4 } from 'uuid';
-import { db } from '../../../../knexfile';
+import { db } from '../../../knexfile';
 import {
   DeploymentRequestDeploymentType,
   DeploymentRequestHubStatus,
@@ -10,25 +10,25 @@ import {
   ServiceConfigurationStatus,
   ServiceDefinitionIdentifier,
   ServiceInstanceCreationStatus,
-} from '../../../__generated__/resolvers-types';
-import { requestContext } from '../../../context/request.context';
-import { OrganizationId } from '../../../model/kanel/public/Organization';
+} from '../../__generated__/resolvers-types';
+import { requestContext } from '../../context/request.context';
+import { OrganizationId } from '../../model/kanel/public/Organization';
 import ServiceInstance, {
   ServiceInstanceId,
-} from '../../../model/kanel/public/ServiceInstance';
-import { SubscriptionId } from '../../../model/kanel/public/Subscription';
-import { securityGuard } from '../../../security/guard';
-import { ErrorCode } from '../../../utils/error/error.code';
-import { FullyQualifiedDeploymentRequest } from '../../deployment/deployment.domain';
-import { loadOrganizationsByUser } from '../../organizations/organizations.domain';
+} from '../../model/kanel/public/ServiceInstance';
+import { SubscriptionId } from '../../model/kanel/public/Subscription';
+import { securityGuard } from '../../security/guard';
+import { ErrorCode } from '../../utils/error/error.code';
+import { FullyQualifiedDeploymentRequest } from '../deployment/deployment.domain';
+import { loadOrganizationsByUser } from '../organizations/organizations.domain';
 import {
   createSubscription,
   loadSubscriptionBy,
   transferSubscriptionToOrganization,
-} from '../../subcription/subscription.domain';
-import { ServiceContractDomain } from '../contract/service-configuration.domain';
+} from '../subcription/subscription.domain';
+import { ServiceConfigurationDomain } from './service-configuration/service-configuration.domain';
 
-import { serviceInstanceDomain } from '../instances/domain';
+import { serviceInstanceDomain } from '../services/instances/domain';
 import { serviceDefinitionIdentifierMappedByPlatformIdentifier } from './registration.mapping';
 
 export type PlatformConfiguration = {
@@ -88,7 +88,7 @@ export const registrationDomain = {
     });
 
     if (configuration) {
-      await ServiceContractDomain.createConfiguration(
+      await ServiceConfigurationDomain.createConfiguration(
         serviceInstanceId,
         configuration
       );
@@ -134,7 +134,7 @@ export const registrationDomain = {
       });
     }
 
-    await ServiceContractDomain.updateConfiguration(serviceInstanceId, {
+    await ServiceConfigurationDomain.updateConfiguration(serviceInstanceId, {
       config: configuration,
       status: ServiceConfigurationStatus.Active,
     });
