+++ apps/portal-api/src/modules/registration/service-configuration/service-configuration.graphql
