+++ apps/portal-api/src/modules/services/service-instance.app.ts
@@ -20,13 +20,13 @@ import { securityGuard } from '../../security/guard';
 import { logApp } from '../../utils/app-logger.util';
 import { ErrorCode } from '../../utils/error/error.code';
 import { NotFoundError } from '../../utils/error/error.util';
+import { PlatformConfiguration } from '../registration/registration.domain';
 import { subscriptionApp } from '../subcription/subscription.app';
 import { loadSubscriptionBy } from '../subcription/subscription.domain';
 import { GenericServiceCapabilityIds } from '../user_service/service-capability/generic_service_capability.const';
 import { UserServiceDomain } from '../user_service/user_service.domain';
 import { uploadNewFile } from './document/document.helper';
 import { Upload } from './document/document.uploads.helper';
-import { PlatformConfiguration } from './registration/registration.domain';
 import {
   grantServiceAccess,
   loadPlatformConfigurationByServiceInstanceId,
