+++ apps/portal-api/src/security/directive-graphql/validators/platform-token-validator.ts
@@ -12,7 +12,7 @@ import {
   loadOrganizationBy,
   organizationDomain,
 } from '../../../modules/organizations/organizations.domain';
-import { ServiceContractDomain } from '../../../modules/services/contract/service-configuration.domain';
+import { ServiceConfigurationDomain } from '../../../modules/registration/service-configuration/service-configuration.domain';
 import { PLATFORM_USER_EMAIL, PLATFORM_USER_UUID } from '../../../portal.const';
 import { logApp } from '../../../utils/app-logger.util';
 
@@ -49,7 +49,7 @@ export const validateActivePlatformToken = async (
   if (!validateExistsToken(req)) return false;
 
   const serviceConfiguration =
-    await ServiceContractDomain.loadConfigurationByPlatformAndToken({
+    await ServiceConfigurationDomain.loadConfigurationByPlatformAndToken({
       platformId: extractPlatformId(req),
       token: extractPlatformToken(req),
     });
@@ -95,7 +95,7 @@ const loadOrganizationFromPlatformIdAndTokenHeaders = async (
   }
 
   const serviceConfiguration =
-    await ServiceContractDomain.loadConfigurationByPlatformAndToken({
+    await ServiceConfigurationDomain.loadConfigurationByPlatformAndToken({
       platformId: extractPlatformId(req),
       token: extractPlatformToken(req),
     });
