+++ apps/portal-api/src/server/graphql-schema.ts
@@ -8,12 +8,12 @@ import serviceGroupResolver from '../modules/deployment/group/service-group.reso
 import ingestManifestResolver from '../modules/ingest-manifest/ingest-manifest.resolver';
 import logResolver from '../modules/log/log.resolver';
 import organizationsResolver from '../modules/organizations/organizations.resolver';
+import registrationResolver from '../modules/registration/registration.resolver';
 import ServiceDefinitionResolver from '../modules/services/definition/service-definition.resolver';
 import vaultResolver from '../modules/services/document/document.resolver';
 import openAEVScenariosResolver from '../modules/services/document/openaev/scenarios/scenarios.resolver';
 import customDashboardsResolver from '../modules/services/document/opencti/custom-dashboards/custom-dashboards.resolver';
 import integrationsResolver from '../modules/services/document/opencti/integrations/integrations.resolver';
-import registrationResolver from '../modules/services/registration/registration.resolver';
 import ServiceInstanceResolver from '../modules/services/services.resolver';
 import settingsResolver from '../modules/settings/settings.resolver';
 import subscriptionsResolver from '../modules/subcription/subscription.resolver';
